refactor: SM-41 Image 컴포넌트 모듈화
